# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/manifests": {
            "get": {
                "tags": [
                    "Manifest"
                ],
                "summary": "获取应用列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Manifest"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "根据 Manifest 创建 Manifest，自动生成名称",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manifest"
                ],
                "summary": "创建 Manifest",
                "parameters": [
                    {
                        "description": "Manifest 数据（branch 必填）",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Manifest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg_api.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/manifests/{id}": {
            "get": {
                "tags": [
                    "Manifest"
                ],
                "summary": "获取应用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Manifest ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Manifest"
                        }
                    }
                }
            },
            "patch": {
                "description": "部分更新 Manifest（仅支持 digest / commit_hash）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manifest"
                ],
                "summary": "Patch Manifest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Manifest ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patch 数据",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.PatchManifestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "获取项目列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Project"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的项目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project"
                ],
                "summary": "创建项目",
                "parameters": [
                    {
                        "description": "Project Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg_api.CreateResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{id}": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "获取项目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Project"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Project"
                ],
                "summary": "更新项目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Project Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Project"
                ],
                "summary": "删除项目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{id}/applications": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "获取项目下的应用列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Application"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_bsonger_devflow-app-service_pkg_model.Application": {
            "type": "object",
            "properties": {
                "active_manifest_id": {
                    "type": "string"
                },
                "active_manifest_name": {
                    "type": "string"
                },
                "config_maps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.ConfigMap"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "envs": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.EnvVar"
                        }
                    }
                },
                "id": {
                    "type": "string"
                },
                "internet": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Internet"
                },
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "project_name": {
                    "type": "string"
                },
                "replica": {
                    "type": "integer"
                },
                "repo_url": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Service"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.ReleaseType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.ConfigMap": {
            "type": "object",
            "properties": {
                "files_path": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "mount_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_bsonger_devflow-app-service_pkg_model.EnvVar": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "github_com_bsonger_devflow-app-service_pkg_model.Internet": {
            "type": "string",
            "enum": [
                "internal",
                "external"
            ],
            "x-enum-varnames": [
                "Internal",
                "External"
            ]
        },
...
        "github_com_bsonger_devflow-app-service_pkg_model.Manifest": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "application_name": {
                    "type": "string"
                },
                "branch": {
                    "type": "string"
                },
                "commit_hash": {
                    "type": "string"
                },
                "config_maps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.ConfigMap"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "digest": {
                    "type": "string"
                },
                "envs": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.EnvVar"
                        }
                    }
                },
                "execution_intent_id": {
                    "type": "string"
                },
                "git_repo": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "internet": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Internet"
                },
                "name": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "replica": {
                    "type": "integer"
                },
                "service": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Service"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.ManifestStatus"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.ManifestStep"
                    }
                },
                "type": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.ReleaseType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.ManifestStatus": {
            "type": "string",
            "enum": [
                "Pending",
                "Running",
                "Succeeded",
                "Failed"
            ],
            "x-enum-varnames": [
                "ManifestPending",
                "ManifestRunning",
                "ManifestSucceeded",
                "ManifestFailed"
            ]
        },
        "github_com_bsonger_devflow-app-service_pkg_model.ManifestStep": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.StepStatus"
                },
                "task_name": {
                    "type": "string"
                },
                "task_run": {
                    "type": "string"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.PatchManifestRequest": {
            "type": "object",
            "properties": {
                "commit_hash": {
                    "type": "string"
                },
                "digest": {
                    "type": "string"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.Port": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "target_port": {
                    "type": "integer"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.Project": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.ProjectStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.ProjectStatus": {
            "type": "string",
            "enum": [
                "active",
                "archived"
            ],
            "x-enum-varnames": [
                "ProjectActive",
                "ProjectArchived"
            ]
        },
        "github_com_bsonger_devflow-app-service_pkg_model.ReleaseType": {
            "type": "string",
            "enum": [
                "normal",
                "canary",
                "blue-green"
            ],
            "x-enum-varnames": [
                "Normal",
                "Canary",
                "BlueGreen"
            ]
        },
        "github_com_bsonger_devflow-app-service_pkg_model.Service": {
            "type": "object",
            "properties": {
                "ports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Port"
                    }
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.StepStatus": {
            "type": "string",
            "enum": [
                "Pending",
                "Running",
                "Succeeded",
                "Failed"
            ],
            "x-enum-varnames": [
                "StepPending",
                "StepRunning",
                "StepSucceeded",
                "StepFailed"
            ]
        },
        "pkg_api.CreateResponse": {
            "type": "object",
            "properties": {
                "execution_intent_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
    }
}